requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
pillow==10.0.1
matplotlib==3.7.2
easyocr==1.7.0
pytesseract==0.3.10
psutil==5.9.6
